MSQ-9 : Implemented "Remember me" login feature
Added a "Remember me" checkbox to the login form.

If the option is checked, a cookie will be created in the browser and stored in the database.

When the user visits the website again, they will be logged in using their local browser cookie if it matches the one stored in the database.

The cookie created contains OS information gathered from the browser's userAgent. This allows for a user to have different login cookies for multiple devices instead of always overwriting the previous cookie associated with their user id.